refactor: Merge Layout Components
