+++ src/widgets/LayoutComponent/index.ts
@@ -1,0 +1,4 @@
+export { Header } from "./ui/Header";
+export { SearchModal } from "./ui/SearchModal";
+export { SimpleLayout } from "./ui/SimpleLayout";
+export { SpacialControl } from "./ui/SpacialControl";
