+++ src/widgets/LayoutComponent/ui/SearchModal/ModalClose.tsx
