+++ src/widgets/LayoutComponent/ui/SimpleLayout.tsx
