+++ src/widgets/LayoutComponent/ui/SpacialControl/SearchBtn.tsx
