--- src/widgets/LayoutHeader/index.ts
@@ -1,2 +1,0 @@
-export { Header } from "./ui/Header";
-
