--- src/widgets/LayoutSpacialControl/index.ts
@@ -1,1 +1,0 @@
-export { SpacialControl } from "./ui/SpacialControl";
